feat: Kruskal Minigame script
Minigame script features the following functionality (reset minigame state, loading a new graph/level and callbacks to aforementioned UnionFind functions).

Correction of description in UnionFind data structure as it alone can find a spanning tree and not an MST. Added a readable flag for the minigame to determine game completion.